pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.17.2",
    "cos-python-sdk-v5>=1.9.39",
    "fastapi[all]>=0.124.4",
    "filelock>=3.20.1",
    "json-repair>=0.54.3",
    "openai>=2.11.0",
    "psycopg2-binary>=2.9.11",
    "pydantic-settings>=2.12.0",
    "redis>=7.1.0",
    "sqlalchemy>=2.0.45",
    "syncpg>=1.1.3",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.2",
]
